[integration tests] move config/data to utils/data
