+++ test_integration/lib/test/realtime/realtime_presence_enter_update_leave.dart
@@ -2,8 +2,8 @@
 import 'package:ably_flutter/ably_flutter.dart';
 import 'package:ably_flutter_example/provisioning.dart';
 
-import '../../config/data.dart';
 import '../../factory/reporter.dart';
+import '../../utils/data.dart';
 
 final logMessages = <List<String>>[];
