+++ test_integration/lib/test/rest/rest_presence_get_test.dart
@@ -1,8 +1,8 @@
 import 'package:ably_flutter/ably_flutter.dart';
 import 'package:ably_flutter_example/provisioning.dart';
 
-import '../../config/data.dart';
 import '../../factory/reporter.dart';
+import '../../utils/data.dart';
 import '../../utils/rest.dart';
 
 final logMessages = <List<String>>[];
